pyproject.toml
[build-system]
requires = ["setuptools>=41", "wheel", "setuptools-git-versioning<2"]
build-backend = "setuptools.build_meta"

# https://setuptools-git-versioning.readthedocs.io/en/stable/install.html
[tool.setuptools-git-versioning]
enabled = true

[project]
name = "pycityproto"
description = "City Proto Generated Python SDK"
readme = "README.md"
dynamic = ["version"]
requires-python = ">=3.8"
dependencies = [
  'importlib-metadata; python_version<"3.8"',
  "grpcio >= 1.60.0, < 2.0.0",
  "protobuf >= 4.24.0, < 5.0.0",
  "types-protobuf >= 4.24.0, < 5.0.0",
]
classifiers = [
  "Programming Language :: Python :: 3",
  "Operating System :: OS Independent",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["pycityproto*"]
